fix(webapp): confirm an OAuth decline before demanding a human
Review found the first cut over-claiming in both directions: it called a
throttled GitHub token dead, and it treated every `onSilentRenew` null as a
lapsed session even though provider hooks collapse transport failures into
null too. Exit 3 now costs a user a login only when that is established.

- Only a 401 is `rejected`. GitHub answers 403 for primary and secondary rate
  limits, so an undifferentiated 403 stays `unknown` — a throttled caller must
  never be sent to a consent window that cannot fix anything.
- The validation call is bounded (`AbortSignal.timeout`), so `--check` always
  answers instead of hanging with no status and no exit code.
- A declined `--renew` runs the upstream check before reporting: confirmed
  REJECTED keeps exit 3 and the "stop and ask a human" remedy; UNKNOWN or no
  hook reports the decline as unconfirmed and names the likely fix without
  claiming a retry is pointless; ACCEPTED says the stored token still works,
  so callers are not blocked and no login is needed.
- A renewal that THREW asks for a retry rather than a consent window.
- `--check` on a refused token points at `--renew` first when the provider can
  renew — a stored refresh token may replace the access token with no human —
  and only exits 3 when no silent renewal exists.
- Expiry wording derives from `expiresAt`; `expired` is a 60s renewal buffer,
  so it now adds "renewal needed" rather than claiming the timestamp passed.
- Document `onValidateToken` and these rules in the adding-slicc-features
  developer skill.

The added help text pushed the eager first-load closure 5.5 kB over its 5 kB
allowance, so `oauth-token` takes the repo's stub + `run.ts` split: the whole
command now loads on first use and the eager graph is 17 kB smaller than the
merge-base.
